requirements.txt
arrow==0.12.0
click==6.7
Flask==0.12.2
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
py-postgresql==1.2.1
python-dateutil==2.6.1
six==1.11.0
Werkzeug==0.14.1
